eventlisteners moved from presenter to views
