+++ src/main.js
@@ -4,7 +4,7 @@ import FooterStatistics from './view/footer-statistics-view.js';
 import FilmsPresenter from './presenter/films-presenter.js';
 import FilmsModel from './model/films-model.js';
 import CommentsModel from './model/comments-model.js';
-import { render } from './render.js';
+import { render } from './framework/render.js';
 
 /**
  * Page variables
